Implement drive input support
This new input has the following syntax:

	message: 'Select a drive'
	type: 'drive'
	name: 'drive'

and triggers a widget that scans the system for available drives, and
prompts the user to select one.

Currently, this widget makes `form.run()` fail with an error if no
available drives were found, but this will soon change once the widget
is extended to auto detect new plugged devices.