Fix `when` properties not taking effect
Before, questions were passed as a whole to `inquirer`, which
understands `when` properties based on the answers to the other passed
questions.

Now, in order to be able to implement the `drive` input, we fold the
questions and pass them in isolation to inquirer, therefore, inquirer
doesn't have enough context anymore to be able to understand `when`
properties.

Solution:

- Check if a question should be asked or not ourselves.
- Since we need to make sure no `when` property arrives to inquirer
(since it can erroneusly be evaluated to false), we change `when` to
`shouldPrompt` internally after parsing the questions.