Implement override option
This option allows the client to override questions in the form with
predefined values and avoid the question to be asked interactively.

For example:

	form.run [
		message: 'What\'s your name?'
		name: 'name'
		type: 'input'
	],
		override:
			name: 'Juan Cruz'
	.then (answers) ->
		console.log(answers)

In the above case, `name` will always equal `Juan Cruz`, and the input
will not even be presented to the user.

The motivation behind this feature is to more easily allow command line
options to override certain asked questions.