chore(package): Update test dependencies
This updates the tests to avoid a dependency causing deprecation warnings
with Node 10. Updating the dependency in question (`gulp-mocha`) doesn't
have the desired effect, as support for on-the-fly compiled CoffeeScript
was dropped.

Due to this the tests are converted to plain old JavaScript, and run with
`mocha` directly, while the `test` gulp task is removed.

Change-Type: patch